requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
requests==2.32.3
sqlalchemy==2.0.32

python-docx==1.1.2
jinja2==3.1.4
rapidfuzz==3.9.6

pypdf==5.1.0
python-multipart==0.0.9
